[WIP] Improve the editor theme [ci skip]
The editor theme now makes use of rounded corners and less borders
to follow modern visual trends.

The default theme's colors were also tweaked to make the blue hue
more subtle (similar to the Arc theme, which was removed as a
consequence). The Alien theme was replaced by a Breeze Dark theme,
which should blend in well with the KDE theme.

PopupMenus (and derivatives such as OptionButton) also now have
a "pressed" style that can be used in projects.